+++ src/index.ts
@@ -2,7 +2,7 @@ export { buildContractClass, VerifyResult } from './contract';
 export { compile, StructEntity, getStructDeclaration, getABIDeclaration, ABIEntity, ABIEntityType, ABI } from './compilerWrapper';
 export {
     bsv, signTx, toHex, getPreimage, num2bin, bin2num, bool2Asm, int2Asm, parseLiteral, bytes2Literal, bytesToHexString, getValidatedHexString, literal2ScryptType, VariableType, literal2Asm,
-    findStructByType, findStructByName, checkStruct, isStructType, isArrayType
+    findStructByType, findStructByName, checkStruct, isStructType, isArrayType, compileContract
 } from './utils';
 export { serializeState, deserializeState, State, STATE_LEN_2BYTES, STATE_LEN_4BYTES } from './serializer';
 export { Int, Bool, Bytes, PrivKey, PubKey, Sig, Ripemd160, Sha1, Sha256, SigHashType, SigHashPreimage, OpCodeType, SingletonParamType, SupportedParamType, ScryptType, ValueType, Struct, StructObject} from './scryptTypes';
