+++ test/helper.ts
@@ -24,40 +24,3 @@ export function newTx(inputSatoshis: number) {
   };
   return new bsv.Transaction().from(utxo);
 }
-
-
-function getCIScryptc(): string | undefined {
-  const scryptc = join(__dirname, '../', getPlatformScryptc());
-  return existsSync(scryptc) ? scryptc : undefined;
-}
-
-export function compileContract(file: string) {
-  console.log(`Compiling contract ${file} ...`);
-
-
-  if(!existsSync(file)) {
-    throw(`file ${file} not exists!`);
-  }
-
-  var argv = require('minimist')(process.argv.slice(2));
-
-  let scryptc = argv.scryptc;
-  if(argv.ci) {
-    scryptc = getCIScryptc();
-  }
-
-  const result = compile(
-    { path: file },
-    { desc: true, debug: true, outputDir: join(__dirname, '../out'),
-		  cmdPrefix: scryptc
-    }
-  );
-
-  if (result.errors.length > 0) {
-    console.log(`Contract ${file} compiling failed with errors:`);
-    console.log(result.errors);
-    throw result.errors;
-  }
-
-  return result;
-}
